requirements.txt
falcon==3.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
